Add files via upload
I created a list to store all the tasks. I used functions to add tasks, view tasks, and delete tasks. A while loop keeps the program running until the user chooses to exit. If-else conditions check the user’s choice and run the correct function. The program shows a menu and performs actions based on the user’s input.